Refactor table head into separate file
